+++ internal/core/spec/errors.go
@@ -24,5 +24,4 @@ var (
 	ErrInvalidParamValue                   = errors.New("invalid parameter value")
 	ErrStepCommandIsEmpty                  = errors.New("step command is empty")
 	ErrStepCommandMustBeArrayOrString      = errors.New("step command must be an array of strings or a string")
-	ErrStepCommandIsRequired               = errors.New("step command is required")
 )